@@ -9,8 +9,8 @@ import ru.DTF98.ewm.category.mapper.CategoryMapper;
 import ru.DTF98.ewm.category.model.Category;
 import ru.DTF98.ewm.category.service.CategoryService;
 
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
 import java.util.List;
 
 @RestController
