@@ -5,6 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import ru.DTF98.ewm.error.BadRequestException;
 import ru.DTF98.ewm.event.args.AdminSearchEventsArgs;
 import ru.DTF98.ewm.event.args.EventWithViewsArgs;
 import ru.DTF98.ewm.event.dto.EventFullDto;
@@ -13,9 +14,9 @@ import ru.DTF98.ewm.event.mapper.EventMapper;
 import ru.DTF98.ewm.event.service.EventService;
 import ru.DTF98.ewm.event.validation.ValidSearchDateInterval;
 
-import javax.validation.Valid;
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -34,13 +35,22 @@ public class EventAdminController {
                                         @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeStart,
                                         @RequestParam(required = false)
                                         @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeEnd,
+                                        @RequestParam(required = false) Double lat,
+                                        @RequestParam(required = false) Double lon,
+                                        @RequestParam(required = false) Double radius,
                                         @RequestParam(required = false) List<Long> users,
                                         @RequestParam(required = false) List<String> states,
                                         @RequestParam(required = false) List<Long> categories,
                                         @RequestParam(defaultValue = "0") @PositiveOrZero Integer from,
                                         @RequestParam(defaultValue = "10") @Positive Integer size) {
+        if (lat != null || lon != null || radius != null) {
+            if (lat == null || lon == null || radius == null) {
+                throw new BadRequestException("Параметры lat, lon и radius должны быть инициализированны вместе " +
+                        "или отсутствовать вообще.");
+            }
+        }
         AdminSearchEventsArgs args = eventMapper.toAdminSearchEventsArgs(rangeStart, rangeEnd, users, states, categories,
-                from, size);
+                lat, lon, radius, from, size);
         log.info("ADMIN: Получение списка всех событий args={}", args);
         List<EventWithViewsArgs> events = eventService.getAllEventsByAdmin(args);
         return eventMapper.toEventFullDto(events);
@@ -51,7 +61,8 @@ public class EventAdminController {
                                    @Valid @RequestBody UpdateEventAdminRequest updateEventAdminRequest) {
         log.info("ADMIN: Обновление события eventId={}, updateEventAdminRequest={}",
                 eventId, updateEventAdminRequest);
-        EventWithViewsArgs event = eventService.updateEventByAdmin(eventId, eventMapper.toEventAdminUpdateArgs(updateEventAdminRequest));
+        EventWithViewsArgs event = eventService.updateEventByAdmin(eventId,
+                eventMapper.toEventAdminUpdateArgs(updateEventAdminRequest));
         return eventMapper.toEventFullDto(event);
     }
 }
