@@ -2,7 +2,7 @@ package ru.DTF98.ewm.event.dto;
 
 import lombok.*;
 
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.Size;
 import java.util.List;
 
 @NoArgsConstructor
