@@ -5,8 +5,8 @@ import lombok.*;
 import ru.DTF98.ewm.event.enums.EventAdminActionState;
 import ru.DTF98.ewm.event.validation.ValidFutureDate;
 
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.Size;
 import java.time.LocalDateTime;
 
 @NoArgsConstructor
