@@ -1,9 +1,9 @@
-package ru.DTF98.ewm.location;
+package ru.DTF98.ewm.event.model;
 
 import lombok.*;
 
-import javax.persistence.*;
-
+import jakarta.persistence.*;
+import java.sql.Timestamp;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -21,9 +21,18 @@ public class Location {
     @Column(name = "name")
     private String name;
 
+    @Column(name = "expiration_date")
+    private Timestamp expirationDate;
+
+    @Column(name = "permanent")
+    private Boolean permanent;
+
     @Column(name = "lat")
     private Double lat;
 
     @Column(name = "lon")
     private Double lon;
+
+    @Column(name = "radius")
+    private Double radius;
 }
