@@ -1,8 +1,8 @@
 package ru.DTF98.ewm.event.validation;
 
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 import java.time.LocalDateTime;
 
 public class ValidFutureDateValidator implements ConstraintValidator<ValidFutureDate, Object> {
@@ -24,7 +24,7 @@ public class ValidFutureDateValidator implements ConstraintValidator<ValidFuture
                 return true;
             }
         } catch (Exception ignore) {
-            // ignore
+            //
         }
         return false;
     }
