@@ -3,7 +3,6 @@ package ru.DTF98.ewm.event.args;
 import lombok.Builder;
 import lombok.Getter;
 import ru.DTF98.ewm.event.enums.EventUserActionState;
-import ru.DTF98.ewm.location.Location;
 
 import java.sql.Timestamp;
 
@@ -20,7 +19,7 @@ public class EventUserUpdateArgs {
 
     private final Timestamp eventDate;
 
-    private final Location location;
+    private final NewLocationArgs location;
 
     private final Boolean requestModeration;
 
