@@ -4,7 +4,6 @@ package ru.DTF98.ewm.event.args;
 import lombok.Builder;
 import lombok.Getter;
 import ru.DTF98.ewm.event.enums.EventState;
-import ru.DTF98.ewm.location.Location;
 
 import java.sql.Timestamp;
 
@@ -21,7 +20,7 @@ public class NewEventArgs {
 
     private final Timestamp eventDate;
 
-    private final Location location;
+    private final NewLocationArgs location;
 
     private final Boolean requestModeration;
 
