@@ -1,20 +1,26 @@
 package ru.DTF98.ewm.event.dto;
 
-
 import lombok.*;
 
-import jakarta.validation.constraints.Positive;
+import java.sql.Timestamp;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Builder
 @ToString
-public class LocationDto {
-    @Positive
+public class LocationFullDto {
     private Long id;
+
+    private String name;
+
+    private Timestamp expirationDate;
+
+    private Boolean permanent;
 
     private Double lat;
 
     private Double lon;
+
+    private Double radius;
 }
