@@ -1,9 +1,13 @@
 package ru.DTF98.ewm.event.repository;
 
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import ru.DTF98.ewm.location.Location;
+import ru.DTF98.ewm.event.model.Location;
+
+import java.util.List;
 
 @Repository
 public interface LocationRepository extends CrudRepository<Location, Long> {
+    List<Location> findAll(Pageable page);
 }
