@@ -1,23 +1,21 @@
 package ru.DTF98.ewm.filter;
 
 
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 import ru.DTF98.stats.client.StatClient;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.time.LocalDateTime;
 
 @Component
 @RequiredArgsConstructor
-@Slf4j
 public class StatHitFilter extends OncePerRequestFilter {
     private final StatClient statsClient;
 
@@ -26,7 +24,6 @@ public class StatHitFilter extends OncePerRequestFilter {
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-
         filterChain.doFilter(request, response);
 
         if (response.getStatus() / 200 == 1) {
